+++ .prettierrc
@@ -1,0 +1,11 @@
+{
+  "printWidth": 100,
+  "tabWidth": 2,
+  "singleQuote": true,
+  "trailingComma": "all",
+  "bracketSpacing": true,
+  "semi": true,
+  "useTabs": true,
+  "arrowParens": "avoid",
+  "endOfLine": "lf"
+}
