+++ app.json
@@ -1,5 +1,5 @@
 {
-  "name": "nextjs",
+  "name": "testjs",
   "options": {
     "allow-unauthenticated": true,
     "memory": "256Mi",
