+++ components/layouts/Header.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Head from 'next/head';
+
+const Header = () => {
+	return (
+		<Head>
+			<title>hi</title>
+			<link rel="icon" href="/favicon.ico" />
+		</Head>
+	);
+};
+
+export default Header;
