+++ components/layouts/MenuContainer.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MenuContainer = () => {
+	return <nav>123</nav>;
+};
+
+export default MenuContainer;
