+++ package.json
@@ -2,11 +2,33 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build"
+    "build": "next build",
+    "start": "next start"
   },
   "dependencies": {
+    "@types/react-transition-group": "^4.4.4",
+    "bootstrap-css-only": "^4.4.1",
     "next": "latest",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-countdown": "^2.3.1",
+    "react-dom": "^17.0.2",
+    "react-html-email": "^3.0.0",
+    "react-modal-video": "^1.2.6",
+    "react-router-hash-link": "^2.4.0",
+    "react-scroll": "^1.8.5",
+    "react-slick": "^0.28.1",
+    "react-tilt": "^0.1.4",
+    "react-transition-group": "^4.4.2",
+    "reactstrap": "^9.0.1",
+    "sass-loader": "^12.4.0",
+    "use-mobile-detect-hook": "^1.0.4",
+    "web-vitals": "^1.0.1"
+  },
+  "devDependencies": {
+    "node-sass": "^7.0.1",
+    "prettier": "^2.5.1",
+    "@types/node": "^17.0.19",
+    "@types/react": "^17.0.39",
+    "typescript": "^4.5.5"
   }
 }
