+++ pages/index.tsx
@@ -1,0 +1,45 @@
+import Head from 'next/head';
+import styles from '../styles/Home.module.css';
+import MenuContainer from '../components/layouts/MenuContainer';
+import Header from '../components/layouts/Header';
+import FooterContainer from '../components/layouts/FooterContainer';
+
+export default function Home() {
+	return (
+		<div className={styles.container}>
+			<main className={styles.main}>
+				<Header />
+				<MenuContainer />
+				<p className={styles.description}>
+					Get started by editing <code className={styles.code}>pages/∂index.js</code>
+				</p>
+
+				<div className={styles.grid}>
+					<a href="https://nextjs.org/docs" className={styles.card}>
+						<h3>Documentation &rarr;</h3>
+						<p>Find in-depth information about Next.js features and API.</p>
+					</a>
+
+					<a href="https://nextjs.org/learn" className={styles.card}>
+						<h3>Learn &rarr;</h3>
+						<p>Learn about Next.js in an interactive course with quizzes!</p>
+					</a>
+
+					<a href="https://github.com/vercel/next.js/tree/canary/examples" className={styles.card}>
+						<h3>Examples &rarr;</h3>
+						<p>Discover and deploy boilerplate example Next.js projects.</p>
+					</a>
+
+					<a
+						href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+						className={styles.card}
+					>
+						<h3>Deploy &rarr;</h3>
+						<p>Instantly deploy your Next.js site to a public URL with Vercel.</p>
+					</a>
+				</div>
+			</main>
+			<FooterContainer />
+		</div>
+	);
+}
