+++ public/index.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    
+    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.png" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate">
+    <meta http-equiv="Pragma" content="no-cache">
+    <meta http-equiv="Expires" content="0">
+    <meta property="og:type" content="website">
+    <meta property="og:url" content="https://greendata.kr">
+    <meta property="og:title" content="한국그린데이터">
+    <meta property="og:image" content="%PUBLIC_URL%/assets/images/meta.png">
+    <meta property="og:description" content="안녕하세요 리조트 AI 에너지 효율화 기업 한국그린데이터입니다.">
+    <meta property="og:site_name" content="한국그린데이터 홈페이지">
+    <meta property="og:locale" content="ko_KR">
+    <link href="https://fonts.googleapis.com/css?family=Poppins:300,400,500,600,700" rel="stylesheet">
+    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+
+     <title>GreenData Korea | Energy Efficiency</title>
+    <base href="/">
+  </head>
+  <body class="landing-page" data-spy="scroll" data-target=".navbar" data-offset="75">
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <!-- Preloader -->
+    <div class="loader-wrapper">
+      <div class="loader"></div>
+    </div>
+    
+    <div id="root"></div>
+    <div id="color">
+      
+    </div>
+    
+  </body>
+</html>
